chore: move test out of example
